nautilus_execution::client

Trait ExecutionClient

pub trait ExecutionClient {

    // Required methods
    fn is_connected(&self) -> bool;
    fn client_id(&self) -> ClientId;
    fn account_id(&self) -> AccountId;
    fn venue(&self) -> Venue;
    fn oms_type(&self) -> OmsType;
    fn get_account(&self) -> Option<AccountAny>;
    fn generate_account_state(
        &self,
        balances: Vec<AccountBalance>,
        margins: Vec<MarginBalance>,
        reported: bool,
        ts_event: UnixNanos,
    ) -> Result<()>;
    fn start(&mut self) -> Result<()>;
    fn stop(&mut self) -> Result<()>;
    fn submit_order(&self, cmd: &SubmitOrder) -> Result<()>;
    fn submit_order_list(&self, cmd: &SubmitOrderList) -> Result<()>;
    fn modify_order(&self, cmd: &ModifyOrder) -> Result<()>;
    fn cancel_order(&self, cmd: &CancelOrder) -> Result<()>;
    fn cancel_all_orders(&self, cmd: &CancelAllOrders) -> Result<()>;
    fn batch_cancel_orders(&self, cmd: &BatchCancelOrders) -> Result<()>;
    fn query_order(&self, cmd: &QueryOrder) -> Result<()>;

}

Required Methods

fn is_connected(&self) -> bool

fn client_id(&self) -> ClientId

fn account_id(&self) -> AccountId

fn venue(&self) -> Venue

fn oms_type(&self) -> OmsType

fn get_account(&self) -> Option<AccountAny>

fn generate_account_state( &self, balances: Vec<AccountBalance>, margins: Vec<MarginBalance>, reported: bool, ts_event: UnixNanos, ) -> Result<()>

Generates and publishes the account state event.

Errors

Returns an error if generating the account state fails.

fn start(&mut self) -> Result<()>

Starts the execution client.

Errors

Returns an error if the client fails to start.

fn stop(&mut self) -> Result<()>

Stops the execution client.

Errors

Returns an error if the client fails to stop.

fn submit_order(&self, cmd: &SubmitOrder) -> Result<()>

Submits a single order command to the execution venue.

Errors

Returns an error if submission fails.

fn submit_order_list(&self, cmd: &SubmitOrderList) -> Result<()>

Submits a list of orders to the execution venue.

Errors

Returns an error if submission fails.

fn modify_order(&self, cmd: &ModifyOrder) -> Result<()>

Modifies an existing order.

Errors

Returns an error if modification fails.

fn cancel_order(&self, cmd: &CancelOrder) -> Result<()>

Cancels a specific order.

Errors

Returns an error if cancellation fails.

fn cancel_all_orders(&self, cmd: &CancelAllOrders) -> Result<()>

Cancels all orders.

Errors

Returns an error if cancellation fails.

fn batch_cancel_orders(&self, cmd: &BatchCancelOrders) -> Result<()>

Cancels a batch of orders.

Errors

Returns an error if batch cancellation fails.

fn query_order(&self, cmd: &QueryOrder) -> Result<()>

Queries the status of an order.

Errors

Returns an error if the query fails.